pet-clinic-web/src/test/java/org/abdul/petclinic/controller/VetRestControllerTest.java
package org.abdul.petclinic.controller;

import org.abdul.petclinic.model.Vet;
import org.abdul.petclinic.service.VetService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.Collections;

import static org.hamcrest.Matchers.hasSize;
import static org.hamcrest.Matchers.is;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ExtendWith(MockitoExtension.class)
class VetRestControllerTest {
    @Mock
    private VetService vetService;
    @InjectMocks
    private VetRestController vetRestController;
    private MockMvc mockMvc;

    @BeforeEach
    void setUp() {
        mockMvc = MockMvcBuilders.standaloneSetup(vetRestController).build();
    }

    @Test
    void getAllVets() throws Exception {
        //given
        Vet vet = Vet.builder().id(1L).firstName("Abdul").build();
        when(vetService.findAll()).thenReturn(Collections.singleton(vet));

        //then
        mockMvc.perform(get("/api/vets")
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", hasSize(1)))
                .andExpect(jsonPath("$[0].id", is(1)))
                .andExpect(jsonPath("$[0].firstName", is("Abdul")));
    }
}